+++ src/tint/writer/glsl/generator_impl.cc
@@ -2177,7 +2177,7 @@ bool GeneratorImpl::EmitLiteral(std::ostream& out, const ast::LiteralExpression*
             } else if (std::isnan(l->value)) {
                 out << "uintBitsToFloat(0x7fc00000u)";
             } else {
-                out << FloatToString(l->value) << "f";
+                out << FloatToString(static_cast<float>(l->value)) << "f";
             }
             return true;
         },
